simple parser error recovery
Whenever an error is reported, a ParsingErrorException is thrown. This
exception is then caught either by the parent block or the root Parse()
function. Synchronisation is then done to find the next statement to
parse. This is done by simply advancing to after the next semicolon, to
the next keyword, or to after the next closed brace.