modules and build system
Created ProjectManager class, that scans the project directory and
builds a module environment based of the file structure. This is then
passed to the parser and type checker. The type checker uses it to get
classes and functions from different modules (files). Classes and
functions in other modules are referred to using "->" arrow syntax.

The "new" keyword was added, which allows for object initialisation. The
"dot" operator was also added, to make it possible to access members
from within classes.